requirements.txt
Flask
pandas
markdown
pykap
gunicorn
google-genai

